move file handling into env.ts, rename fileConfig to namedConfigs.
